requirements.txt
fastapi
uvicorn
tensorflow
pillow
pandas